pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pagans"
dynamic = ["version"]
description = "😅 Prompts Aligned to Guidelines and Normalization System - Optimize prompts for OpenAI, Anthropic, and Google models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
keywords = [
    "ai",
    "llm",
    "prompt",
    "optimization",
    "openai",
    "anthropic",
    "claude",
    "google",
    "gemini",
    "openrouter",
    "machine-learning",
    "nlp",
    "artificial-intelligence"
]
authors = [
    {name = "PAGANS Team", email = "contact@pagans.dev"}
]
maintainers = [
    {name = "PAGANS Team", email = "contact@pagans.dev"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.28.1",
    "jinja2>=3.1.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[project.urls]
Homepage = "https://github.com/abubakarsiddik31/pagans"
Documentation = "https://pagans.readthedocs.io/"
Repository = "https://github.com/abubakarsiddik31/pagans"
Issues = "https://github.com/abubakarsiddik31/pagans/issues"
Changelog = "https://github.com/abubakarsiddik31/pagans/blob/main/CHANGELOG.md"
"Source Code" = "https://github.com/abubakarsiddik31/pagans"
"Download" = "https://pypi.org/project/pagans/"
"Bug Reports" = "https://github.com/abubakarsiddik31/pagans/issues"

[project.scripts]
pagans = "main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/pagans/_version.py"
root = "."
version_file = "src/pagans/__version__.py"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["S101", "S104", "COM812", "ISC001"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.setuptools.package-data]
"pagans" = ["optimizer_prompts/templates/**/*.jinja"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src/pagans"]
omit = [
    "*/tests/*",
    "*/test_*",
    "src/pagans/__version__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
